Number guessing game
One of the simplest two-player games is “Guess the number” based on PYTHON .The first player thinks of a secret number in some known range. After each guess, the first player answers either “Higher”, “Lower” or “Correct!” depending on whether the secret number is higher, lower or equal to the guess. Score of both players will be calculated and we will get winner.